backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
braintree==4.34.0
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
frozenlist==1.6.0
gunicorn==20.1.0
idna==3.10
multidict==6.4.3
phonenumbers==9.0.3
phonenumberslite==9.0.3
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
twilio==9.6.1
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.20.0